add dummy application example
An example dummy using rawDataTable. Uses basics sorting, filtering and pagination.

Note: to avoid having to convert field names (firstName <-> first-name), capitalization is avoided for field names.

Data is generated locally in the adapter (so it generates ember-data objects), but with a 300ms delay to simulate some waiting for a request.