add if/else for handling icon-links in example raw-data-table, making it more obvious this should be supported
The readme mentions that icons are supported if the icon name is passed, but nothing is actually done with the data. An example if/else is set for raw-data-table, so the user has an obvious place to change if they want to support this.

Doing this differently (e.g. just a comment explaining what to add to support icons) might be better.