Allow passing components directly as custom fields via `@customFieldComponents`.
Uses the same logical flow for `@customFields`. Instead, `@customFields` could just accept both a string or an object.

The component is provided in `fields` (and by way of having the same meaning, `firstColumn` and `otherColumns`).
The user should never do something with this info, as the ember-data implementation will handle the rendering via the custom component.

This avoids having to program if/else structures when having custom rendering for some attributes. Instead, the correct component can be passed.
It is important to note that `<td></td>` is not added and the customComponent should add this.
This means you can't directly pass a specific component used in other parts of the app (e.g. formattedDate.hbs), as `<td></td>` needs to be added.
However, this problem is easily solved, without any extra files, by using the template syntax and passing `<template><td><FormatedDate @date=@cell.value /></td></template>`.
This way @customFieldComponents can use the same logic @customFields does (which also does not set `td` tag, but expects the passed block to use it).