add `@selectionProperty` to support row selection for non-ember-data objects
The default selection will compare based on the whole object (reference), which will not work for paginated results that are POJOs (e.g. mu-search). For this, a comparison should be made on the unique ID, like id for ember-data or uuid for mu-search.
This will still work for ember-data objects when specifying correctly with "id", it is just not needed.

avoids using emberArray's findBy, as this will be deprecated in ember 5.

Should be a plain function helper after updating beyond ember 4.5

This is the only place ember-composable-helpers was used, which means this dependency can be removed. This package got deprecated too, so removing it is positive too.