refactor away `@wrapper` + refactor rowLink(Model)/onRowClicked
`@wrapper` gave the selection status, but this was not used anywhere (as the row gives this in the hash anyway). It also increased complexity in usage, also for the consumer, as the item was not as easily accessible anymore and some extra knowledge about this wrapper was needed.

Instead, the item and related rowLink(Model) are passed down separately. `rowLinkModel` is now calculated at the row level (instead of providing the function at content-body level).

`@onRowClicked` is now changed so the consumer only gets access to `rowClicked`, which also handles row links and this function is moved to the row level. This is renamed from `onRowClicked` as it is not a callback, but an actual function (action) to call when a row is clicked.

wrapping could instead be done for rowLink, rowLinkModel and rowClicked function, but this seemed needlessly complex to avoid passing these three values.