allow passing arrays/objects directly to `@fields`, `@links`, `@customHeaders` and `@customFields`
This adds some features and changes to already existing features:
- `@fields` can now accept an array too. This array can contain the usual options-string, or an object containing the options. If an object is passed, things like the default of the specification will still work (e.g. here the label property will be set if it is missing from the object).

- `@links` has the same change, where an array can be provided with potential objects that already contain the keys as mentioned.

- `@customHeader/Field`: `@customFieldComponents` already existed to pass components directly, instead of having to use the block syntax. In this commit, this is removed and instead components can directly be passed to `@customField` by passing an object (hash) instead of a string/array. Same for `@customHeaders`.
To still allow a combination of both using the named block and passing components to use for rendering, the syntax is used that in the object, a key can have an "empty" value (like an empty string).
This might not be super clear: maybe an explicit string ("block") would work better? This can be easily added.