add more demo routes for multiple types of example tables
this also changes the order that an item is added to the selection, so newly clicked items are always *at the end* of the array. This gives the user a way to know in what order a person clicked the items (if they would ever need it).

The two helpers could be local helpers (in future ember)
allcaps.hbs should just accept a value instead of also adding <td>, but that would only be easy to use in `@customFields` with <template> syntax, which is not yet supported here.