rewrite old tests for new raw-data-table
As the user should now creat their own "raw-data-table" component, and it also contains quite a bit of functionality (that a user should take over), the tests are changed to integration tests that test from a rendered RawDataTable with set properties, instead of only the specific components (which don't contain any UI with the new implementation).

These are still limited, but already give a good headstart.